Fix beberapa bug

+++ controller/Jabatan.js
@@ -1,8 +1,4 @@
-const {
-	uploadFile,
-	createFolder,
-	deleteFile,
-} = require("../service/google-api/googleapi");
+const { uploadFile, deleteFile } = require("../service/google-api/googleapi");
 const Pegawai = require("../models/Pegawai");
 
 module.exports = {
@@ -10,11 +6,15 @@ module.exports = {
 		const { nip } = req.params;
 		try {
 			const pegawai = await Pegawai.findOne({ nip });
-			const file = await uploadFile(
-				req.file,
-				`${req.body.filename} - ${pegawai.nama_lengkap}`,
-				pegawai.folderId.jabatan
-			);
+			const file =
+				req.file !== undefined
+					? await uploadFile(
+							req.file,
+							`${req.body.filename} - ${pegawai.nama_lengkap}`,
+							pegawai.folderId.jabatan
+					  )
+					: "-";
+
 			const newData = {
 				jabatan: req.body.jabatan,
 				divisi: req.body.divisi,
@@ -22,7 +22,7 @@ module.exports = {
 				periode: req.body.periode,
 				filename: req.body.filename,
 				keterangan: req.body.keterangan,
-				fileId: file.id,
+				fileId: file.id || file,
 				folderId: pegawai.folderId.jabatan,
 			};
 			await Pegawai.findOneAndUpdate({ nip }, { $push: { jabatan: newData } });
@@ -41,27 +41,9 @@ module.exports = {
 	},
 
 	findAll: async (req, res) => {
-    const { nip } = req.params;
-    try {
-      const results = await Pegawai.findOne({ nip });
-      return res.ok({
-        success: true,
-        message: "Berhasil Mendapatkan Data",
-        content: results.jabatan,
-      });
-    } catch (err) {
-      return res.error({
-        success: false,
-        message: "Gagal Mendapatkan Data",
-        content: err,
-      });
-    }
-  },
-
-	findById: async (req, res) => {
-		const { nip, id } = req.params;
+		const { nip } = req.params;
 		try {
-			const results = await Pegawai.findOne({ nip, "jabatan._id": id });
+			const results = await Pegawai.findOne({ nip });
 			return res.ok({
 				success: true,
 				message: "Berhasil Mendapatkan Data",
@@ -76,26 +58,52 @@ module.exports = {
 		}
 	},
 
+	findById: async (req, res) => {
+		const { nip, id } = req.params;
+		try {
+			const results = await Pegawai.findOne({ nip, "jabatan._id": id });
+			return res.ok({
+				success: true,
+				message: "Berhasil Mendapatkan Data",
+				content: results.jabatan.filter((el) => el._id == id),
+			});
+		} catch (err) {
+			return res.error({
+				success: false,
+				message: "Gagal Mendapatkan Data",
+				content: err,
+			});
+		}
+	},
+
 	update: async (req, res) => {
 		const { nip, id } = req.params;
 		try {
-			const pegawai = await Pegawai.findOne({ nip });
-			const data = pegawai.jabatan.filter((d) => d._id == id)[0];
-			await deleteFile(data.fileId);
-			const fileJabatan = await uploadFile(
-				req.file,
-				`${req.body.filename} - ${pegawai.nama_lengkap}`,
-				data.folderId
-			);
+			const data = await Pegawai.findOne({ nip, "jabatan._id": id });
+
+			if (data.jabatan[0].fileId !== "-") {
+				await deleteFile(data.jabatan[0].fileId);
+			}
+
+			const fileJabatan =
+				req.file !== undefined
+					? await uploadFile(
+							req.file,
+							`${req.filename} - ${data.nama_lengkap}`,
+							data.folderId.jabatan
+					  )
+					: "-";
+
 			const updateData = {
 				jabatan: req.body.jabatan,
 				divisi: req.body.divisi,
-				aktif: req.body.aktif || pegawai.aktif,
+				aktif: req.body.aktif,
 				periode: req.body.periode,
-				filename: req.body.filename,
+				filename: req.file === undefined ? fileJabatan : req.body.filename,
 				keterangan: req.body.keterangan,
-				fileId: fileJabatan.id,
+				fileId: fileJabatan.id || fileJabatan,
 			};
+
 			await Pegawai.updateOne(
 				{ nip, "jabatan._id": id },
 				{
@@ -128,7 +136,10 @@ module.exports = {
 		const { nip, id } = req.params;
 		try {
 			const data = await Pegawai.findOne({ nip, "jabatan._id": id });
-			await deleteFile(data.jabatan[0].fileId);
+			if (data.jabatan[0].fileId !== "-") {
+				await deleteFile(data.jabatan[0].fileId);
+			}
+
 			await Pegawai.updateOne(
 				{ nip, "jabatan._id": id },
 				{
